Address comments (#24872)
