+++ modules/phase_field/test/tests/free_energy_material/tests
@@ -75,6 +75,15 @@
       detail = "implemented using the compile time derivatives system using automatic differentiation and applied to a split Cahn-Hilliard "
                "system"
     []
+    [ctd-split-ad-jac]
+      type = 'PetscJacobianTester'
+      input = 'MathCTDFreeEnergy_split.i'
+      cli_args = 'Executioner/num_steps=1 Materials/free_energy/type=ADMathCTDFreeEnergy Kernels/cres/type=ADSplitCHParsed'
+      detail = "implemented using the compile time derivatives system using automatic differentiation, with correct dual number derivative entries"
+      ratio_tol = 1e-8
+      difference_tol = 2e-6
+      heavy = true
+    []
   []
 
   [RegularSolutionFreeEnergy]
